+++ Assets/Scripts/Helper/Image/Unmount.sh
@@ -17,13 +17,31 @@ Unmount_All() {
 }
 
 Unmount() {
-  local image_path=$1
-  local mounted_dir=$2
+  local mounted_dir=$1
+  local image_path=$2
 
-  UI "Unmounting: $(basename $mounted_dir)"
-  sudo umount $mounted_dir
-  rm -rf $mounted_dir
-  e2fsck -fa "$image_path" >/dev/null
-  resize2fs -M "$image_path" 2>&1 | grep -v resize2fs >/dev/null
-  UI "d"
+  if Is_Extracted "$mounted_dir"; then
+    return 0
+  fi
+
+  if mountpoint -q "$mounted_dir"; then
+    if [ ! -d "$mounted_dir" ]; then
+      UI "!!Unmount: $mounted_dir not a valid directory."
+      exit 1
+    fi
+
+    UI "Unmounting: $(basename $mounted_dir)"
+    sudo umount $mounted_dir
+    rm -rf $mounted_dir
+    if [ -n "$image_path" ]; then
+      if [[ ! "$image_path" =~ \.img$ ]]; then
+        UI "!!Unmount: $image_path must be a .img"
+        exit 1
+      fi
+      e2fsck -fa "$image_path" >/dev/null
+      resize2fs -M "$image_path" 2>&1 | grep -v resize2fs >/dev/null
+    fi
+    UI "d"
+  fi
+
 }
